% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pb.R
\name{clean_case}
\alias{clean_case}
\title{Clean Case Numbers}
\usage{
clean_case(x)
}
\arguments{
\item{x}{is a character vector containing case numbers.}
}
\value{
Character vector of standardized case numbers.
}
\description{
Useful for correcting manually entered case numbers that may be abbreviated or contain non-standard methods for handling null values such as "No #"
}
\examples{
clean_case("PB123") # PB000123
clean_case("None") # NA
clean_case("5678") # PB005678
}
\seealso{
clean_x, clean_id
}

