% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pb.R
\name{clean_pb}
\alias{clean_pb}
\title{Clean PB Numbers}
\usage{
clean_pb(x)
}
\arguments{
\item{x}{is a character vector containing PB numbers}
}
\value{
Character vector of standardized PB numbers. NA values are introduced where there appears to be non-X characters present.
}
\description{
Useful for correcting manually entered PB numbers that may be abbreviated or contain non-standard methods for handling null values such as "No #"
}
\examples{
clean_pb("PB123") # PB000123
clean_x("None") # NA
clean_x("5678") # PB005678
}
\seealso{
clean_x, clean_id
}

