% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_id.R
\name{obtain_id}
\alias{obtain_id}
\title{Obtain ID Numbers}
\usage{
obtain_id(obtain = list(), using = list(), value = "")
}
\arguments{
\item{obtain}{character. The desired identifier. Could be either "X" or "CII".}

\item{using}{character. Name of the known identifier. Could be "X", "CASE", or "CII".}

\item{value}{character vector. The actual value(s) of the known identifier.}
}
\value{
List of length 1 containing the desired unique identifier
}
\description{
Obtains an unknown identifier given some known identifier.
}
\examples{
obtain_id(obtain = "X", using = "CASE", "PB021665")
}
\seealso{
clean_case, clean_x, clean_id
}

