% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_x.R
\name{clean_x}
\alias{clean_x}
\title{Clean X Numbers}
\usage{
clean_x(x)
}
\arguments{
\item{x}{is a character vector containing X numbers}
}
\value{
Character vector of standardized X numbers. NA values are introduced where there appears to be non-X characters present.
}
\description{
Useful for correcting manually entered X-numbers that may be abbreviated or contain non-standard methods for handling null values such as "No #"
}
\examples{
clean_x("X6789") # X00006789
clean_x("No #") # NA
clean_x("1234") # X00001234
}
\seealso{
clean_pb, clean_id
}

